\name{enzyme}
\alias{enzyme}
\docType{data}
\title{
Enzymatic activity in the blood
}
\description{
These data correspond to representing the metabolism of carcinogenic substances among 245 unrelated individuals.
}

\usage{data(enzyme)}
\format{\code{enzyme} is a data frame with 245 cases (rows).
}
\details{
For more information about dataset see Bechtel et al. (1993).}

\source{
Bechtel, Y., Bonaiti-Pellie, C., Poisson, N., Magnette, J. and Bechtel, P. (1993). A population and family study of n-acetyltransferase using caffeine urinary metabolites. Clinical Pharmacology and Therapeutics, 54, 134-141.
}
%\examples{
%\dontrun{

%}
%}


\keyword{datasets}
