\name{ozone}
\alias{ozone}
\docType{data}
\title{
Daily ozone level measurements
}
\description{
These data correspond to daily ozone level measurements (in \eqn{ppb=ppm x 1000}) in New York in May-September, 1973, from the New York State Department of Conservation.
}

\usage{data(ozone)}
\format{\code{ozone} is a data frame with 116 cases (rows).
}
\details{
For a complete description of various distributions applied to data concentration of air pollutants see Gokhale and Khare (2004).}

\source{
Leiva, V., Barros, M., Paula, G. e Sanhueza, A. (2007). Generalized BirnbaumSaunders distribution applied to air pollutant concentration. Environmetrics, 19, 235-249.

Nadarajah, S. (2007). A truncated inverted beta distribution with application to air
pollution data. Stoch. Environ. Res. Risk. Assess., 22, 285-289.

Gokhale, S. e Khare, M. (2004) A review of deterministic, stochastic and hybrid vehicular exhaust emission models International. J. Transp. Manag., 2, 59-74.
}
%\examples{
%\dontrun{

%}
%}
\keyword{datasets}
